modification des untiés en mètre et seconde
Former-commit-id: 00ccedc621b5cfd221f04c97a72284b51fdc2231

+++ src/structures/vehicle.h
@@ -8,7 +8,7 @@
 typedef struct Vehicle {
     char* name; // Nom du véhicule
     int fastCharge; // Capacité de charge rapide en km/h
-    float range; // Autonomie
+    int range; // Autonomie
 } Vehicle;
 
 Vehicle* readJSONvehicles(char* filename, int* n);
